Add files via upload
Commented out portions of make_mm.py that was creating dummy mulitpole and POL PTS sections, as these are not necessary for prot.efp in Qchem (May be needed for GAMESS?). 

Fixed cut_qm.py to NOT remove dipole, quadrupole, and octupole sections of atoms that are"second-neighbors" to QM region; previously these were removed along with POL PTS-- now only POL PTS are removed.